move getMaxVols function to predicates.go and change the params of NewMaxPDVolumeCountPredicate funcs
